requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
python-dotenv>=1.0.0
openai>=1.54.0
httpx>=0.28.0
PyGithub>=2.5.0
aiofiles>=24.1.0
gitpython>=3.1.40
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
